% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype.R
\name{auto_detect_tissue_type}
\alias{auto_detect_tissue_type}
\title{Auto detect tissue: Guess tissue type from seurat data}
\usage{
auto_detect_tissue_type(
  path_to_db_file,
  seuratObject,
  scaled,
  assay = "RNA",
  ...
)
}
\arguments{
\item{path_to_db_file}{db path}

\item{seuratObject}{seurat object}
}
\description{
Auto detect tissue: Guess tissue type from seurat data
}
