% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CompareMarkersByPublicDB}
\alias{CompareMarkersByPublicDB}
\title{CompareMarkers}
\usage{
CompareMarkersByPublicDB(
  data = NULL,
  fdb = NULL,
  col_ct = "cellName",
  col_gene = "geneSymbol"
)
}
\arguments{
\item{data}{object}

\item{col_ct}{cell column name}

\item{col_gene}{gene column name}
}
\value{
data frame
}
\description{
CompareMarkers
}
