% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FilterMarkers_SeuratForm}
\alias{FilterMarkers_SeuratForm}
\title{Filter markers for Seurat form marker table}
\usage{
FilterMarkers_SeuratForm(
  data = NULL,
  gene_col = "gene",
  p_val_col = "p_val_adj",
  log2fc_col = "avg_log2FC",
  pct1_col = "pct.1",
  pval = 1e-04,
  log2fc = 0.585,
  pct1 = 0.3
)
}
\arguments{
\item{data}{object}
}
\value{
data frame
}
\description{
Filter markers for Seurat form marker table
}
