% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype.R
\name{gene_sets_prepare}
\alias{gene_sets_prepare}
\title{gene_sets_prepare}
\usage{
gene_sets_prepare(path_to_db_file, cell_type)
}
\arguments{
\item{path_to_db_file}{db path}

\item{cell_type}{cell type tissue}
}
\description{
gene_sets_prepare
}
