% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_util.R
\name{RunFindAllMarkers}
\alias{RunFindAllMarkers}
\title{Gene expression markers for all group}
\usage{
RunFindAllMarkers(
  obj = NULL,
  assay_use = "SCT",
  idents = "seurat_clusters",
  logfc = 0.25,
  min_pct = 0.1,
  min_diff_pct = 0,
  recorrect_umi = TRUE
)
}
\arguments{
\item{obj}{object}

\item{assay_use}{assay}
}
\value{
data frame
}
\description{
Gene expression markers for all group
}
