% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype.R
\name{sctype_score}
\alias{sctype_score}
\title{sctype_score}
\usage{
sctype_score(
  scRNAseqData,
  scaled = !0,
  gs,
  gs2 = NULL,
  gene_names_to_uppercase = !0,
  ...
)
}
\arguments{
\item{scRNAseqData}{matrix}
}
\description{
sctype_score
}
