% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype.R
\name{ScTypeAnnotation}
\alias{ScTypeAnnotation}
\title{CellType Anno SCType}
\usage{
ScTypeAnnotation(obj, assay, reduction, f_marker_db, tissue, out_path)
}
\arguments{
\item{obj}{seurat object}
}
\description{
CellType Anno SCType
}
