% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_util.R
\name{SeuratAddScorePlus}
\alias{SeuratAddScorePlus}
\title{Seurat Add Score}
\usage{
SeuratAddScorePlus(
  obj = NULL,
  assay = "SCT",
  slot = "data",
  name = "",
  features = "",
  fdata = "",
  form = "seurat"
)
}
\arguments{
\item{obj}{seurat object}

\item{name}{score title}

\item{features}{gene vectors}
}
\value{
seurat object
}
\description{
Seurat Add Score
}
