% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_util.R
\name{SeuratAnchors}
\alias{SeuratAnchors}
\title{Seruat Anchors}
\usage{
SeuratAnchors(
  que_obj = NULL,
  ref_obj = NULL,
  nfeatures = 5000,
  norm_method = "LogNormalize",
  reduction = "cca",
  dim1 = 1,
  dim2 = 30,
  weight_reduction = "pca"
)
}
\arguments{
\item{que_obj}{seurat object of query}

\item{ref_obj}{seurat object of reference}
}
\description{
Seruat Anchors
}
