% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SeuratMarkerForm_to_scType}
\alias{SeuratMarkerForm_to_scType}
\title{Convert the Seurat marker table format to the scType format}
\usage{
SeuratMarkerForm_to_scType(
  data = NULL,
  group_col = "cluster",
  gene_col = "gene",
  tissueType = "Brain",
  source_name = "",
  outfile = "sctypeForm.xlsx"
)
}
\arguments{
\item{data}{object}
}
\value{
data frame
}
\description{
Convert the Seurat marker table format to the scType format
}
